Generate Plan working successfully

+++ user/models.py
@@ -13,8 +13,8 @@ To achieve this
 class UserData(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, null=True, blank=True, related_name="user_data")
     user_district = models.CharField(max_length=100, null=True)
-    curr_lat = models.DecimalField(max_digits=11, decimal_places=8, null=True, blank=True)
-    curr_log = models.DecimalField(max_digits=11, decimal_places=8, null=True, blank=True)
+    curr_lat = models.FloatField(null=True, blank=True)
+    curr_log = models.FloatField(null=True, blank=True)
 
     def __str__(self):
         return self.user_district
